+++ context/app/static/js/components/detailPage/related-entities/RelatedEntitiesTabs/RelatedEntitiesTabs.jsx
@@ -13,13 +13,18 @@ function RelatedEntitiesTabs({ entities, openIndex, setOpenIndex, ariaLabel, ren
     <>
       <StyledTabs value={openIndex} onChange={handleChange} aria-label={ariaLabel}>
         {entities.map((entity, i) => (
-          <Tab label={`${entity.tabLabel} (${entity.data.length})`} index={i} key={entity.tabLabel} />
+          <Tab
+            label={`${entity.tabLabel} (${entity.data.length})`}
+            index={i}
+            key={entity.tabLabel}
+            data-testid={`${entity.tabLabel.toLowerCase()}-tab`}
+          />
         ))}
       </StyledTabs>
       {entities.map(({ tabLabel, data, entityType: tableEntityType, columns }, i) => (
-        <StyledTabPanel value={openIndex} index={i} key={tabLabel}>
+        <StyledTabPanel value={openIndex} index={i} key={tabLabel} data-testid={`${tabLabel.toLowerCase()}-panel`}>
           {data.length > 0 ? (
-            <RelatedEntitiesTable columns={columns} entities={data} />
+            <RelatedEntitiesTable columns={columns} entities={data} entityType={tableEntityType.toLowerCase()} />
           ) : (
             <StyledAlert severity="warning">{renderWarningMessage(tableEntityType)}</StyledAlert>
           )}