+++ context/app/static/js/pages/Donor/Donor.jsx
@@ -68,7 +68,7 @@ function DonorDetail() {
           description={description}
           group_name={group_name}
         />
-        {shouldDisplaySection.metadata && <MetadataTable />}
+        {shouldDisplaySection.metadata && <MetadataTable metadata={mapped_metadata} hubmap_id={hubmap_id} />}
         <DerivedEntitiesSection sectionId="derived" />
         <ProvSection />
         {shouldDisplaySection.protocols && <Protocol protocol_url={protocol_url} />}