+++ context/app/static/js/shared-styles/panels/PanelListLandingPage/PanelListLandingPage.jsx
@@ -6,13 +6,13 @@ import { PageWrapper, StyledDescription } from './style';
 function PanelListLandingPage({ title, subtitle, description, children }) {
   return (
     <PageWrapper>
-      <Typography variant="h2" component="h1">
+      <Typography variant="h2" component="h1" data-testid="landing-page-title">
         {title}
       </Typography>
-      <Typography variant="subtitle1" color="primary">
+      <Typography variant="subtitle1" color="primary" data-testid="landing-page-subtitle">
         {subtitle}
       </Typography>
-      {description && <StyledDescription>{description}</StyledDescription>}
+      {description && <StyledDescription data-testid="landing-page-description">{description}</StyledDescription>}
       {children}
     </PageWrapper>
   );