+++ context/app/markdown/CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.73.0 - 2023-06-20
+
+- Fix donor metadata table display.
+- Set up template for Biomarkers Gene Page. 
+- Built gene data API.
+- Built gene id API. 
+- Rendered summary on summary section from gene data API.
+- Fix type error on donor pages with no derived samples.
+- Fix errors with query-portal-logs script.
+- Release and QA instructions have been migrated to Confluence.
+
+
 ## v0.72.4 - 2023-06-15
 
 - Built two custom hooks to call third party API in order to retrieve the common name to render it next to the gene symbol.